Move DLL loading code from client to libkb for kbfs
