+++ go/libkb/saferdllloading_dummy.go
@@ -3,7 +3,7 @@
 // Copyright 2016 Keybase, Inc. All rights reserved. Use of
 // this source code is governed by the included BSD license.
 
-package client
+package libkb
 
 // SaferDLLLoading dummy for platforms not needing this.
 func SaferDLLLoading() error { return nil }
