+++ go/libkb/saferdllloading_windows.go
@@ -1,7 +1,7 @@
 // Copyright 2016 Keybase, Inc. All rights reserved. Use of
 // this source code is governed by the included BSD license.
 
-package client
+package libkb
 
 import (
 	"errors"
